Interpretation of Elizabeth's image
I really loved the somber pathos that was evident throughout Elizabeth's project. And I will try to embody that same pathos for my interpretation. It seems apparant that her images of the train and message reply bubble speaks to the viewer's sense of longing and hesitation. I will attempt a similar mood, with a dog that is waiting for its owner, and a slow clock.